+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### v1.6.0
 * Balance: Rain is now uncommon instead of rare
+* Reworked Severed Torchhead
 
 ### v1.5.6
 * Fix: Deflated Dodgeball now only triggers during combat